+++ src/dawn/native/CompilationMessages.cpp
@@ -140,7 +140,7 @@ void OwnedCompilationMessages::ClearMessages() {
 }
 
 const WGPUCompilationInfo* OwnedCompilationMessages::GetCompilationInfo() {
-    mCompilationInfo.messageCount = mMessages.size();
+    mCompilationInfo.messageCount = static_cast<uint32_t>(mMessages.size());
     mCompilationInfo.messages = mMessages.data();
 
     // Ensure every message points at the correct message string. Cannot do this earlier, since
